+++ glomar_gridding/variogram.py
@@ -7,6 +7,7 @@ matrices.
 """
 
 from dataclasses import dataclass
+from abc import ABC, abstractmethod
 from typing import Literal
 import numpy as np
 import xarray as xr
@@ -15,9 +16,10 @@ from scipy.special import gamma, kv
 
 
 @dataclass()
-class Variogram:
-    """Place holder"""
+class Variogram(ABC):
+    """Generic Variogram Class - defines the abstract class"""
 
+    @abstractmethod
     def fit(
         self, distance_matrix: np.ndarray | xr.DataArray
     ) -> np.ndarray | xr.DataArray:
@@ -225,7 +227,7 @@ class MaternVariogram(Variogram):
     nugget : float | np.ndarray
         The value of the independent variable at distance 0
     effective_range : float | np.ndarray | None
-        Effective Range, this is the lag where 95% of ths sill are exceeded.
+        Effective Range, this is the lag where 95% of the sill are exceeded.
         This is not the range parameter, which is defined as r/3 if nu < 0.5 or
         nu > 10, otherwise r/2 (where r is the effective range). One of
         effective_range and range must be set.